Get the connections between cart, lineitems and products almost working. Need to see why carts are loosing their items

+++ src/main/java/com/baldmountain/depot/DbSeed.java
@@ -50,26 +50,38 @@ public class DbSeed {
         mongoService = MongoService.create(vertx, config);
         mongoService.start();
 
-        mongoService.dropCollection("products", event1 -> {
-            if (event1.succeeded()) {
-                LinkedList<Product> products = new LinkedList<>();
-                products.add(new Product("CoffeeScript", "<p>\n" +
-                        "        CoffeeScript is JavaScript done right. It provides all of JavaScript's " +
-                        "functionality wrapped in a cleaner, more succinct syntax. In the first " +
-                        "book on this exciting new language, CoffeeScript guru Trevor Burnham " +
-                        "shows you how to hold onto all the power and flexibility of JavaScript " +
-                        "while writing clearer, cleaner, and safer code.</p>", "/images/cs.jpg", new BigDecimal(36.00).setScale(2, RoundingMode.CEILING)));
-                products.add(new Product("Programming Ruby 1.9 & 2.0", "<p>Ruby is the fastest growing and most exciting dynamic language " +
-                        "out there. If you need to get working programs delivered fast, " +
-                        "you should add Ruby to your toolbox. </p>", "/images/ruby.jpg", new BigDecimal(49.95).setScale(2, RoundingMode.CEILING)));
-                products.add(new Product("Rails Test Prescriptions", "<p><em>Rails Test Prescriptions</em> is a comprehensive guide to testing " +
-                        "Rails applications, covering Test-Driven Development from both a theoretical perspective (why to test) and from a practical perspective " +
-                        "(how to test effectively). It covers the core Rails testing tools and procedures for Rails 2 and Rails 3, and introduces popular add-ons, " +
-                        "including Cucumber, Shoulda, Machinist, Mocha, and Rcov.</p>", "/images/rtp.jpg", new BigDecimal(34.95).setScale(2, RoundingMode.CEILING)));
-                saveAProduct(products);
-            } else {
-                setDone("Trouble dropping products");
-            }
+        mongoService.dropCollection("line_items", res1 -> {
+           if (res1.succeeded()) {
+               mongoService.dropCollection("carts", res2 -> {
+                   if (res2.succeeded()) {
+                       mongoService.dropCollection("products", event1 -> {
+                           if (event1.succeeded()) {
+                               LinkedList<Product> products = new LinkedList<>();
+                               products.add(new Product("CoffeeScript", "<p>\n" +
+                                       "        CoffeeScript is JavaScript done right. It provides all of JavaScript's " +
+                                       "functionality wrapped in a cleaner, more succinct syntax. In the first " +
+                                       "book on this exciting new language, CoffeeScript guru Trevor Burnham " +
+                                       "shows you how to hold onto all the power and flexibility of JavaScript " +
+                                       "while writing clearer, cleaner, and safer code.</p>", "/images/cs.jpg", new BigDecimal(36.00).setScale(2, RoundingMode.CEILING)));
+                               products.add(new Product("Programming Ruby 1.9 & 2.0", "<p>Ruby is the fastest growing and most exciting dynamic language " +
+                                       "out there. If you need to get working programs delivered fast, " +
+                                       "you should add Ruby to your toolbox. </p>", "/images/ruby.jpg", new BigDecimal(49.95).setScale(2, RoundingMode.CEILING)));
+                               products.add(new Product("Rails Test Prescriptions", "<p><em>Rails Test Prescriptions</em> is a comprehensive guide to testing " +
+                                       "Rails applications, covering Test-Driven Development from both a theoretical perspective (why to test) and from a practical perspective " +
+                                       "(how to test effectively). It covers the core Rails testing tools and procedures for Rails 2 and Rails 3, and introduces popular add-ons, " +
+                                       "including Cucumber, Shoulda, Machinist, Mocha, and Rcov.</p>", "/images/rtp.jpg", new BigDecimal(34.95).setScale(2, RoundingMode.CEILING)));
+                               saveAProduct(products);
+                           } else {
+                               setDone("Trouble dropping products");
+                           }
+                       });
+                   } else {
+                       setDone("Trouble dropping carts");
+                   }
+               });
+           } else {
+               setDone("Trouble dropping line_items");
+           }
         });
 
         try {
